Refactor crypto modules : separate encryption from key exchange - Extract ChaCha20-Poly1305 encryption logic into [src/crypto/chacha20.rs]); - Implement [MessageCipher] trait in [chacha20.rs]; - Move nonce cache logic, key ratcheting, and key wrapping functions to [chacha20.rs]; - Remove encryption, decryption, and nonce logic from [src/crypto/x25519.rs], keeping only [X25519KeyExchange]; - Update [fish11_decryptmsg.rs] to use `crypto::chacha20` and the new [is_nonce_replay] / [mark_nonce_seen] helpers; - Update [fish11_encryptmsg.rs] to use the `crypto::chacha20` namespace; - Update [fish11_exchangekey.rs] to use the [KeyExchange] trait and `is_socket_connected` utility; - Fix various imports and type usages across the DLL interface. - Cargo fmtp --all
