+++ fish_11_dll/src/crypto/mod.rs
@@ -1,5 +1,25 @@
+use std::any::Any;
+
+/// Trait for message encryption and decryption
+pub trait MessageCipher: Any + Send + Sync {
+    /// Encrypt a message
+    fn encrypt(&self, key: &[u8], message: &str, recipient: Option<&str>, associated_data: Option<&[u8]>) -> Result<String>;
+    
+    /// Decrypt a message
+    fn decrypt(&self, key: &[u8], encrypted_data: &str, associated_data: Option<&[u8]>) -> Result<String>;
+    
+    /// Generate a symmetric key
+    fn generate_key(&self) -> Result<Vec<u8>>;
+
+    /// Upcast to Any
+    fn as_any(&self) -> &dyn Any;
+}
+
 pub mod x25519;
 pub use self::x25519::*;
+
+pub mod chacha20;
+pub use self::chacha20::*;
 
 use crate::error::Result;
